gw: rename nmigen to amaranth
